refactor(search_playground): updated useLLMsModels to useQuery
Refactored the useLLMsModels hook to be a useQuery implementation that
re-uses the load connectors query via the query client. This allows the
result to be cached and additionally will allow calling the LMsModels
query with the query client for loading models with a saved playground.
