+++ x-pack/solutions/search/plugins/search_playground/public/hooks/use_llms_models.ts
@@ -5,14 +5,17 @@
  * 2.0.
  */
 
+import { type QueryClient, useQuery, useQueryClient } from '@tanstack/react-query';
+import { HttpSetup } from '@kbn/core/public';
 import { i18n } from '@kbn/i18n';
-import { useMemo } from 'react';
 import { SERVICE_PROVIDERS } from '@kbn/inference-endpoint-ui-common';
+
 import type { PlaygroundConnector, InferenceActionConnector, ActionConnector } from '../types';
 import { LLMs } from '../../common/types';
 import { LLMModel } from '../types';
-import { useLoadConnectors } from './use_load_connectors';
 import { MODELS } from '../../common/models';
+import { useKibana } from './use_kibana';
+import { LOAD_CONNECTORS_QUERY_KEY, LoadConnectorsQuery } from './use_load_connectors';
 
 const isInferenceActionConnector = (
   connector: ActionConnector
@@ -96,52 +99,66 @@ const mapLlmToModels: Record<
   },
 };
 
+export const LLMS_QUERY_KEY = ['search-playground', 'llms-models'];
+
+export const LLMsQuery =
+  (http: HttpSetup, client: QueryClient) => async (): Promise<LLMModel[]> => {
+    const connectors = await client.fetchQuery<PlaygroundConnector[]>({
+      queryKey: LOAD_CONNECTORS_QUERY_KEY,
+      queryFn: LoadConnectorsQuery(http),
+      retry: false,
+    });
+
+    const mapConnectorTypeToCount = connectors.reduce<Partial<Record<LLMs, number>>>(
+      (result, connector) => {
+        result[connector.type] = (result[connector.type] || 0) + 1;
+        return result;
+      },
+      {}
+    );
+
+    const models = connectors.reduce<LLMModel[]>((result, connector) => {
+      const connectorType = connector.type as LLMs;
+      const llmParams = mapLlmToModels[connectorType];
+
+      if (!llmParams) {
+        return result;
+      }
+
+      const showConnectorName = Number(mapConnectorTypeToCount?.[connectorType]) > 1;
+
+      llmParams
+        .getModels(connector.name, false)
+        .map(({ label, value, promptTokenLimit }) => ({
+          id: connector?.id + label,
+          name: label,
+          value,
+          connectorType: connector.type,
+          connectorName: connector.name,
+          showConnectorName,
+          icon: typeof llmParams.icon === 'function' ? llmParams.icon(connector) : llmParams.icon,
+          disabled: !connector,
+          connectorId: connector.id,
+          promptTokenLimit,
+        }))
+        .forEach((model) => result.push(model));
+
+      return result;
+    }, []);
+
+    return models;
+  };
+
 export const useLLMsModels = (): LLMModel[] => {
-  const { data: connectors } = useLoadConnectors();
+  const client = useQueryClient();
+  const {
+    services: { http },
+  } = useKibana();
 
-  const mapConnectorTypeToCount = useMemo(
-    () =>
-      connectors?.reduce<Partial<Record<LLMs, number>>>(
-        (result, connector) => ({
-          ...result,
-          [connector.type]: (result[connector.type as LLMs] || 0) + 1,
-        }),
-        {}
-      ),
-    [connectors]
-  );
+  const { data } = useQuery(LLMS_QUERY_KEY, LLMsQuery(http, client), {
+    keepPreviousData: true,
+    retry: false,
+  });
 
-  return useMemo(
-    () =>
-      connectors?.reduce<LLMModel[]>((result, connector) => {
-        const connectorType = connector.type as LLMs;
-        const llmParams = mapLlmToModels[connectorType];
-
-        if (!llmParams) {
-          return result;
-        }
-
-        const showConnectorName = Number(mapConnectorTypeToCount?.[connectorType]) > 1;
-
-        return [
-          ...result,
-          ...llmParams
-            .getModels(connector.name, false)
-            .map(({ label, value, promptTokenLimit }) => ({
-              id: connector?.id + label,
-              name: label,
-              value,
-              connectorType: connector.type,
-              connectorName: connector.name,
-              showConnectorName,
-              icon:
-                typeof llmParams.icon === 'function' ? llmParams.icon(connector) : llmParams.icon,
-              disabled: !connector,
-              connectorId: connector.id,
-              promptTokenLimit,
-            })),
-        ];
-      }, []) || [],
-    [connectors, mapConnectorTypeToCount]
-  );
+  return data || [];
 };
