+++ x-pack/solutions/search/plugins/search_playground/public/hooks/use_load_connectors.ts
@@ -8,41 +8,39 @@
 import type { UseQueryResult } from '@tanstack/react-query';
 import { useQuery } from '@tanstack/react-query';
 import { loadAllActions as loadConnectors } from '@kbn/triggers-actions-ui-plugin/public/common/constants';
-import type { IHttpFetchError, ResponseErrorBody } from '@kbn/core-http-browser';
+import type { HttpSetup, IHttpFetchError, ResponseErrorBody } from '@kbn/core-http-browser';
 import { i18n } from '@kbn/i18n';
 import { useKibana } from './use_kibana';
 import { type PlaygroundConnector } from '../types';
 import { parsePlaygroundConnectors } from '../utils/playground_connectors';
 
-const QUERY_KEY = ['search-playground, load-connectors'];
+export const LOAD_CONNECTORS_QUERY_KEY = ['search-playground, load-connectors'];
+
+export const LoadConnectorsQuery = (http: HttpSetup) => async () => {
+  const queryResult = await loadConnectors({ http });
+  return parsePlaygroundConnectors(queryResult, http);
+};
 
 export const useLoadConnectors = (): UseQueryResult<PlaygroundConnector[], IHttpFetchError> => {
   const {
     services: { http, notifications },
   } = useKibana();
 
-  return useQuery(
-    QUERY_KEY,
-    async () => {
-      const queryResult = await loadConnectors({ http });
-      return parsePlaygroundConnectors(queryResult, http);
+  return useQuery(LOAD_CONNECTORS_QUERY_KEY, LoadConnectorsQuery(http), {
+    retry: false,
+    keepPreviousData: true,
+    onError: (error: IHttpFetchError<ResponseErrorBody>) => {
+      if (error.name !== 'AbortError') {
+        notifications?.toasts?.addError(
+          error.body && error.body.message ? new Error(error.body.message) : error,
+          {
+            title: i18n.translate('xpack.searchPlayground.loadConnectorsError', {
+              defaultMessage:
+                'Error loading connectors. Please check your configuration and try again.',
+            }),
+          }
+        );
+      }
     },
-    {
-      retry: false,
-      keepPreviousData: true,
-      onError: (error: IHttpFetchError<ResponseErrorBody>) => {
-        if (error.name !== 'AbortError') {
-          notifications?.toasts?.addError(
-            error.body && error.body.message ? new Error(error.body.message) : error,
-            {
-              title: i18n.translate('xpack.searchPlayground.loadConnectorsError', {
-                defaultMessage:
-                  'Error loading connectors. Please check your configuration and try again.',
-              }),
-            }
-          );
-        }
-      },
-    }
-  );
+  });
 };
